dev: use MySQL 8.0

+++ dev/run-mysql.sh
@@ -3,13 +3,18 @@
 set -exu
 
 options=(--rm -p3306:3306)
+options+=(-eMYSQL_ALLOW_EMPTY_PASSWORD=yes)
 if [ $# -ge 1 ]; then
   db_base_dir=$1
-  rm -rf ${db_base_dir}
+  if [ -e ${db_base_dir} ]; then
+    rm -rf ${db_base_dir}
+    if [ -e ${db_base_dir} ]; then
+      sudo -H rm -rf ${db_base_dir}
+    fi
+  fi
   mkdir -p ${db_base_dir}/mysql
-  mkdir -p ${db_base_dir}/log
-  chmod +t ${db_base_dir}/log
-  chmod go+wx ${db_base_dir}/log
+  mkdir -p ${db_base_dir}/log/mysql
+  chmod -R go+wx ${db_base_dir}/log
   options+=("-v${db_base_dir}/mysql:/var/lib/mysql")
   options+=("-v${db_base_dir}/log:/var/log")
 fi
@@ -22,4 +27,4 @@ max_allowed_packet = 256M
 MY_CNF
 options+=("-v${db_conf_dir}:/etc/my.cnf.d")
 
-docker run "${options[@]}" groonga/mroonga:latest
+docker run "${options[@]}" groonga/mroonga:mysql-8.0-latest
